Sort of semi works
Def not done yet, however, we now sort of have something that almost
does what we desire. It uses breeze gtk dark and then goes through qt
style plugins to set it to breeze dark for qt too, and we use qt5ct and
qt6ct and nwg for some finishing touches (like fonts size). We probably
want to: remove the declarative themeing aspect and instead just install
themes and then just set with gtk directly,
so that nix and qtct/nwg are not fighting for control